+++ test/DailyTaskTest/GetCoinBalance.cs
@@ -18,7 +18,7 @@ namespace DailyTaskTest
             {
                 var dailyTaskAppService = scope.ServiceProvider.GetRequiredService<ICoinDomainService>();
 
-                int number = dailyTaskAppService.GetCoinBalance();
+                decimal number = dailyTaskAppService.GetCoinBalance();
                 Assert.True(number >= 0);
             }
         }
